Move generate_pairwise_instances to util.py related to learners
